requirements.txt
gradio
pandas
snscrape
praw
newspaper3k
nltk
matplotlib
wordcloud
transformers
torch
beautifulsoup4
seaborn